Extracted the HTML format logic to the HtmlFormatter
This allows to format an HTML string without having to use an HtmlTextView in the layout.
Also this allows to customize the html before passing it to the TextView, for example one could change all BulletSpans with its own implementation.
HtmlFormatterBuilder parameter uses the builder pattern.

Removed unused HtmlTagHandler#mTextPaint
Replace HtmlResImageGetter#container TextView with Context